+++ src/core/align-content/index.ts
@@ -1,4 +1,4 @@
-import { GlobalCssKeyword } from '../shared'
+import { GlobalCssKeyword, serializeAtomicValue } from '../shared'
 
 type BaseLinePosition =
   | ['first', 'baseline']
@@ -47,7 +47,10 @@ export type AlignContentDeclaration = {
 
 export const serializeAlignContentPropValue = (
   x: AlignContent | GlobalCssKeyword
-): string => (typeof x === 'string' ? x : `${x[0]} ${x[1]}`)
+): string =>
+  Array.isArray(x)
+    ? `${serializeAtomicValue(x[0])} ${serializeAtomicValue(x[1])}`
+    : serializeAtomicValue(x)
 
 export const serializeAlignContent = (
   x: AlignContent | GlobalCssKeyword