+++ src/core/background-color/index.test.ts
@@ -1,5 +1,5 @@
 import { style } from '../../index'
-import { rgb } from '../color'
+import { rgb } from '../shared'
 
 test('backgroundColor: Color<RGB>', () => {
   const received = style({