+++ src/core/min-height/index.test.ts
@@ -16,7 +16,7 @@ test('minHeight: WidthCalculation', () => {
     minHeight: cmulti(cdiv(per(100), 7), 3),
   })
   const expected = {
-    minHeight: 'calc((100% / 7) * 3)',
+    minHeight: 'calc(calc(100% / 7) * 3)',
   }
   expect(received).toEqual(expected)
 })