+++ src/core/shared/types/index.ts
@@ -11,6 +11,8 @@ export * from './timing-function'
 export * from './calc'
 export * from './decibel'
 export * from './generated-types'
+export * from './color'
+export * from './env'
 
 export const globalCssKeywords = <const>['inherit', 'initial', 'unset']