+++ src/core/shared/types/position.ts
@@ -1,12 +1,21 @@
-import { getData, LengthPercentage, NAMESPACE, RBType } from './shared'
-import { serializeLengthPercentage } from './shared'
+import { WidthCalculation } from './calc'
+import {
+  getData,
+  LengthPercentage,
+  NAMESPACE,
+  RBType,
+  serializeAtomicValue,
+} from './shared'
 
 /**
  * A value of this type defines an (x, y) coordinate.
  * @added 0.2.7
  */
 export interface Position
-  extends RBType<'Position', [LengthPercentage, LengthPercentage]> {}
+  extends RBType<
+    'Position',
+    [LengthPercentage | WidthCalculation, LengthPercentage | WidthCalculation]
+  > {}
 
 /**
  * @category Value constructor
@@ -18,12 +27,13 @@ export function pos(x: LengthPercentage, y: LengthPercentage): Position {
       type: 'Position',
       data: [x, y],
       valueConstructor: pos,
+      serializer: serializePosition,
     },
   }
 }
 
 export const serializePosition = (x: Position): string =>
   getData(x).reduce((acc, val, idx) => {
-    const serializedVal = serializeLengthPercentage(val)
+    const serializedVal = serializeAtomicValue(val)
     return acc + serializedVal + (idx === pos.length - 1 ? '' : ' ')
   }, '')