+++ src/core/shared/types/shared.ts
@@ -36,8 +36,15 @@ export const NAMESPACE = 'npmjs.com/package/rosebox'
 
 export interface RBTypeVal<A, B> {
   type: A
-  valueConstructor: Function
   data: B
+  valueConstructor: Function
+  serializer: Function
+}
+
+export const serializeAtomicValue = (x: string | number | RBType): string => {
+  if (typeof x === 'number') return `${x}`
+  if (typeof x === 'string') return x
+  return x[NAMESPACE].serializer(x)
 }
 
 export type RBType<A = string, B = any> = {
@@ -48,3 +55,4 @@ export const getData = <A, B>(x: RBType<A, B>) => x[NAMESPACE].data
 export const getValConstructor = <A, B>(x: RBType<A, B>) =>
   x[NAMESPACE]?.valueConstructor
 export const getTypeName = (x: any): string => x[NAMESPACE]?.type ?? ''
+export const getSerializer = (x: RBType): Function => x[NAMESPACE].serializer